backend/inventory/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
orjson==3.11.6
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.22
python-ulid==1.1.0
PyYAML==6.0.3
redis==5.3.1
redis-om==0.3.5
requests==2.32.5
rich==14.3.1
rich-toolkit==0.17.2
rignore==0.7.6
sentry-sdk==2.51.0
shellingham==1.5.4
starlette==0.50.0
tomli==2.4.0
typer==0.21.1
types-cffi==1.17.0.20250915
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==80.10.0.20260124
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
